feat: 시간 변환 함수 util로 분리 #79
